Fix joystick member function call that was missing

+++ XdevLDemos/joystick_demo/joystick_demo.cpp
@@ -45,14 +45,16 @@ int main(int argc, char **argv) {
 	//
 	xdl::IPXdevLJoystickServer joystickServer = xdl::createModule<xdl::IPXdevLJoystickServer>(core, xdl::XdevLModuleName("XdevLJoystickServer"), xdl::XdevLID("MyJoystickServer"));
 	if(nullptr == joystickServer) {
-	xdl::destroyCore(core);
+		xdl::destroyCore(core);
 		return -1;
 	}
 
 	// And use that info to connect to the joystick device we want.
 	//
-	xdl::XdevLJoystickDeviceInfo info = joystickServer->getJoystickInfo(xdl::XdevLJoystickId::JOYSTICK_DEFAULT);
-	joystick->create(info);
+	if(joystick->create(joystickServer, xdl::XdevLJoystickId::JOYSTICK_DEFAULT) != xdl::ERR_OK) {
+		xdl::destroyCore(core);
+		return -1;
+	}
 
 	//
 	// Now we use delegates that will help us to managed events. When a specific button or axis is used
@@ -71,7 +73,7 @@ int main(int argc, char **argv) {
 
 	// Start the main loop.
 	for(;;) {
-	core->update();
+		core->update();
 		xdl::sleep(0.001);
 	}
